**What this PR does**:

For the servicegraphs processor:
- Gates this feature behind the EnableExtraUninstrumentedServicesLabels flag in the Config
- If enabled, adds new labels to `traces_service_graph_request_*`:
  - `virtual_node` with a value of `client` or `server`, so the un-instrumented service's node can be identified.
  - `client|server_db|messaging_system` with the corresponding resource attribute value, to surface and identify the service and the node.

**Which issue(s) this PR fixes**:
Fixes #3461 

**Checklist**
- [x] Tests updated
- [x] Documentation added
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`