@t00mas 2024-04-12
> a few initial questions. other thoughts:
> 
> * what occurs if there are conflicts between these new values and configured span dimensions?
> * what do you think about pooling edges in [the store](https://github.com/grafana/tempo/blob/main/modules/generator/processor/servicegraphs/store/store.go)?

Thanks for reviewing Joe.

I've added some testing to cover "conflicts" with manually added `Dimensions`.

Could you be more specific about the edge pooling? Happy to address that if you think there's an issue there. (Sorry, I'm just new to Tempo in general)