@t00mas 2024-04-18
> I wonder if we should make this a per-tenant override instead. If this feature is behind a global config parameter, it makes me think that we should just add it to the default generator without a flag. What's the use case for enabling or disabling it globally?

It actually makes more sense as a per-tenant override instead, yes.

I've also discussing with Joe and I will:
- Scope-down the PR just to add the new `virtual_node` label (the others can be added with a combination of Dimensions and Prefixes)
- Pool edges in the store for performance reasons