@t00mas 2024-04-22
These are some benchmarks about `Edge` reuse with pooling:

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/tempo/modules/generator/processor/servicegraphs
```

|                   | 20_runs_bench.txt | 20_runs_bench_with_pooling.txt      |
| ----------------- | ----------------- | ----------------------------------- |
|                   | sec/op            | sec/op     vs base                  |
| ServiceGraphs-11  | 21.33µ ± 0%       | 21.63µ ± 2%  +1.41% (p=0.006 n=20)  |
|                   | B/op              | B/op      vs base                   |
| ServiceGraphs-11  | 24.89Ki ± 0%      | 24.06Ki ± 0%  -3.35% (p=0.000 n=20) |
|                   | allocs/op         | allocs/op   vs base                 |
| ServiceGraphs-11  | 450.0 ± 0%        | 443.0 ± 0%  -1.56% (p=0.000 n=20)   |

There's probably better numbers to be had at scale, since these are limited to the testdata size.